requirements.txt
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.12.0
pillow>=8.3.0
numpy>=1.20.0
matplotlib>=3.4.0
ipywidgets>=7.6.0
requests>=2.25.0
scikit-learn>=1.0.0
ipython>=7.0.0
